Combine Store and Events components
