--- src/components/onboarding/Store/style.module.scss
@@ -1,32 +1,0 @@
-@use 'src/styles/vars.scss' as vars;
-
-.page {
-  display: flex;
-  flex-direction: column;
-  padding: 2rem;
-
-  .items {
-    display: flex;
-    gap: 1rem;
-    justify-content: space-between;
-    margin-top: 2rem;
-
-    > * {
-      flex: 1;
-    }
-
-    @media (max-width: vars.$breakpoint-lg) {
-      justify-content: center;
-
-      > * {
-        flex: unset;
-      }
-    }
-  }
-
-  @media (max-width: vars.$breakpoint-lg) {
-    .desktopOnly {
-      display: none;
-    }
-  }
-}
