--- src/components/onboarding/Store/style.module.scss.d.ts
@@ -1,11 +1,0 @@
-export type Styles = {
-  desktopOnly: string;
-  items: string;
-  page: string;
-};
-
-export type ClassNames = keyof Styles;
-
-declare const styles: Styles;
-
-export default styles;
